Fix edge case where hops can extend past frequency range.
ceil()ing the count of hops can cause a weird hop count to be rounded up, which can cause the maximum center frequency to exceed the range.
If the range was the device range, this will then cause the last hop to raise an error in the driver, crashing the program.
This fix prevents this.